+++ src/roseingrave/_input_files.py
@@ -73,6 +73,22 @@ def read_template(path=None, strict=False):
     for level, level_values in raw_values.items():
         if level == "validation":
             continue
+        # DEPRECATED: REMOVE IN v1.0.0
+        if level == "masterSpreadsheet":
+            # for backward compatibility, do not fail on warning
+            logger.warning(
+                'Deprecated key "masterSpreadsheet": '
+                'use "summarySpreadsheet" instead'
+            )
+            if "summarySpreadsheet" in raw_values:
+                # they have an extra key, so just skip this value
+                logger.warning(
+                    'Using values of "summarySpreadsheet" key: ',
+                    'values of "masterSpreadsheet" will be ignored',
+                )
+                continue
+            # rename the key so that callers will have the right key
+            level = "summarySpreadsheet"
         if level not in values:
             warning = True
             logger.warning('unknown key "{}"', level)
@@ -115,7 +131,7 @@ def read_template(path=None, strict=False):
 
     # validate values
     # public access options
-    for ss in ("masterSpreadsheet", "volunteerSpreadsheet"):
+    for ss in ("summarySpreadsheet", "volunteerSpreadsheet"):
         value = values[ss]["publicAccess"]
         if value not in PUBLIC_ACCESS_OPTIONS:
             warning = True